requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
sentence-transformers
transformers
torch
unstructured[pdf]
pdf2image
pdfminer.six